Fixed filecheck bugs

+++ scripts/install/mdbfiles.sh
@@ -7,14 +7,16 @@ function InstallMusicDBFiles {
     local MDBUSER="$3"
     local MDBGROUP="$4"
 
-    _ExpectingUser  rsync
+    _ExpectingTool  rsync
     _ExpectingUser  $MDBUSER
     _ExpectingGroup $MDBGROUP
 
     echo -e -n "\e[1;34mInstalling MusicDB files to \e[0;36m$SERVERDIR\e[1;34m: "
 
     WSCONFIG=$SERVERDIR/webui/config.js
-    cp "$WSCONFIG" "/tmp/mdbwebuicfg.bak"
+    if [ -f "$WSCONFIG" ] ; then
+        cp "$WSCONFIG" "/tmp/mdbwebuicfg.bak"
+    fi
 
     rsync -uav  \
         --exclude 'tmp/' \
@@ -35,7 +37,9 @@ function InstallMusicDBFiles {
     cp $SOURCEDIR/scripts/helper/*.sh "$SERVERDIR/."
     set -f
 
-    mv "/tmp/mdbwebuicfg.bak" "$WSCONFIG"
+    if [ -f "/tmp/mdbwebuicfg.bak" ] ; then
+        mv "/tmp/mdbwebuicfg.bak" "$WSCONFIG"
+    fi
     chown -R $MDBUSER:$MDBGROUP $SERVERDIR
 
     echo -e "\e[1;32mdone\e[0m"
